requirements.txt
fastapi
uvicorn[standard]
pymongo
dnspython
python-dotenv
pydantic
email-validator
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pandas
openpyxl
reportlab
python-barcode
Pillow
cryptography
certifi
truststore
